requirements.txt
fastapi
uvicorn
httpx
python-dotenv
beautifulsoup4
playwright
pandas
pdfplumber
pillow
pytesseract
google-generativeai
